requirements.txt
flask~=1.1.2
pysimplemodel~=2.4.3
requests~=2.23.0
bs4~=0.0.1
twilio~=6.42.0
sklearn~=0.0
numpy~=1.18.5
pandas~=1.0.4
matplotlib~=3.2.1
statsmodels~=0.11.1
beautifulsoup4~=4.9.1
soupsieve~=2.0.1
PyJWT~=1.7.1
pip~=19.3.1
wheel~=0.33.6
pytz~=2020.1
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
setuptools~=42.0.2
six~=1.15.0
patsy~=0.5.1
scipy~=1.4.1
MarkupSafe~=1.1.1
joblib~=0.15.1
python-dateutil~=2.8.1
certifi~=2020.4.5.2
chardet~=3.0.4
scikit-learn~=0.23.1
threadpoolctl~=2.1.0
urllib3~=1.25.9
idna~=2.9
cycler~=0.10.0
pyparsing~=2.4.7
kiwisolver~=1.2.0